Refactor leaderboard table and empty state components
